streamlit_ppt_summary/requirements.txt
streamlit
python-pptx
pandas
pyyaml
groq
pytest
python-dotenv